Tried running on testnet but VRFv1 is not working
I'm getting a gas estimation error which I have no idea why. The gas estimation error pops up when I call the end lottery function and in that function, my guess is that the requestRandomness function function is causing the issue. I''m gonna try re-coding Lottery.sol with VRFv2 to see if it helps